move lib/pyitlib.py to hfs/
