+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
     rev: 4.0.1
     hooks:
     -   id: flake8
-        args: ["--max-line-length=90", "--exclude=hfs/lib/,*/__init__.py", "--extend-ignore=E741,W503,W605,E501"]
+        args: ["--max-line-length=90", "--exclude=hfs/pyitlib.py,*/__init__.py", "--extend-ignore=E741,W503,W605,E501"]
         files: ^hfs/
 
   - repo: local
