+++ hfs/pyitlib.py
